# swagger.yaml
openapi: 3.0.0
paths:
...
  /auth/register:
    post:
      operationId: AuthController_register
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterDto'
      responses:
        '201':
          description: Пользователь успешно зарегистрирован.
      tags: &ref_1
        - Аутентификация
  /auth/login:
    post:
      operationId: AuthController_login
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginDto'
      responses:
        '200':
          description: Успешный вход в систему.
      tags: *ref_1
...
info:
  title: API документация
  description: Описание API
  version: '1.0'
  contact: {}
...
servers: []
components:
  schemas:
    RegisterDto:
      type: object
      properties:
        email:
          type: string
          description: Адрес электронной почты пользователя
        password:
          type: string
          description: Пароль пользователя
          minLength: 6
      required:
        - email
        - password
    LoginDto:
      type: object
      properties:
        email:
          type: string
          description: Электронная почта пользователя
        password:
          type: string
          description: Пароль пользователя
      required:
        - email
        - password